feat: 인증 시스템 기반 구조 구현
- AuthContext 및 AuthProvider 구현
- AuthTypes 타입 정의
- useAuth 커스텀 훅 구현
- ProtectedRoute HOC 구현
