+++ src/contexts/AuthContext.tsx
@@ -1,76 +1,5 @@
-import React, { createContext, useContext, useState, useEffect } from 'react';
-import type { ReactNode } from 'react';
-
-interface User {
-    id: string;
-    email: string;
-    name: string;
-}
-
-interface AuthContextType {
-    user: User | null;
-    login: (email: string, password: string) => Promise<boolean>;
-    logout: () => void;
-    isLoading: boolean;
-}
+import { createContext } from 'react';
+import type { AuthContextType } from './AuthTypes';
 
 const AuthContext = createContext<AuthContextType | undefined>(undefined);
-
-export const useAuth = () => {
-    const context = useContext(AuthContext);
-    if (context === undefined) {
-        throw new Error('useAuth must be used within an AuthProvider');
-    }
-    return context;
-};
-
-interface AuthProviderProps {
-    children: ReactNode;
-}
-
-export const AuthProvider: React.FC<AuthProviderProps> = ({ children }) => {
-    const [user, setUser] = useState<User | null>(null);
-    const [isLoading, setIsLoading] = useState(true);
-
-    useEffect(() => {
-
-        const savedUser = localStorage.getItem('user');
-        if (savedUser) {
-            setUser(JSON.parse(savedUser));
-        }
-        setIsLoading(false);
-    }, []);
-
-    const login = async (email: string, password: string): Promise<boolean> => {
-
-        if (email && password) {
-            const mockUser: User = {
-                id: '1',
-                email,
-                name: email.split('@')[0],
-            };
-            setUser(mockUser);
-            localStorage.setItem('user', JSON.stringify(mockUser));
-            return true;
-        }
-        return false;
-    };
-
-    const logout = () => {
-        setUser(null);
-        localStorage.removeItem('user');
-    };
-
-    const value = {
-        user,
-        login,
-        logout,
-        isLoading,
-    };
-
-    return (
-        <AuthContext.Provider value={value}>
-            {children}
-        </AuthContext.Provider>
-    );
-}; 
+export default AuthContext;
