+++ src/hoc/ProtectedRoute.tsx
@@ -1,0 +1,22 @@
+import { Navigate, useLocation } from 'react-router-dom';
+import useAuth from '@/contexts/useAuth';
+import type { ReactNode } from 'react';
+
+interface ProtectedRouteProps {
+    children: ReactNode;
+}
+
+export default function ProtectedRoute({ children }: ProtectedRouteProps) {
+    const { user, isLoading } = useAuth();
+    const location = useLocation();
+
+    if (isLoading) {
+        return <div>로딩 중...</div>;
+    }
+
+    if (!user) {
+        return <Navigate to="/login" state={{ from: location.pathname }} replace />;
+    }
+
+    return <>{children}</>;
+}
